Fixed profile search and queries, and setup API base
All MySQL queries that return user profiles have been changed to support the 'primary email' feature; that is, instead of returning both the login email and alternate email, it only returns the alternate one if not null, or the login one otherwise. Additionally, only the primary email is searched for when doing a keyword search.

The basics of a RESTful API have been set up as well, under /client/api.php. The idea is to have all the API functions in one file if possible. Various functions could be called by visiting /client/api.php?<function_name>, and checking the GET parameters.