Added Logger, changed Profile to require confirmation code when creating
Added Logger.php from the IEFDF project; it has been used in a few places, but still needs to be implemented fully.

Modified Profile.php to keep track of a 'state' variable, set to 'View', 'Create', 'CreatePending', 'Edit', or 'EditPending'. This makes it simpler to decide which views to render.

Changed profile creation to require a confirmation code, same as when editing. The basic logic is there, but it still needs to be fully implemented.